requirements.txt
absl-py==1.0.0
astunparse==1.6.3
attrs==21.2.0
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.3
cloudpickle==2.0.0
cma==3.1.0
contextlib2==21.6.0
cycler==0.11.0
Deprecated==1.2.13
fast-and-efficient==1.0.0
filelock==3.4.0
filterpy==1.4.5
flatbuffers==2.0
fonttools==4.28.2
gast==0.4.0
google-auth==2.3.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.42.0
gym==0.21.0
h5py==3.6.0
idna==3.3
importlib-metadata==4.8.2
importlib-resources==5.4.0
jsonschema==4.2.1
keras==2.7.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
libclang==12.0.0
Markdown==3.3.6
matplotlib==3.5.0
ml-collections==0.1.0
msgpack==1.0.2
numpy==1.21.4
oauthlib==3.1.1
opt-einsum==3.3.0
packaging==21.3
Pillow==8.4.0
protobuf==3.19.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybullet==3.2.0
pyparsing==3.0.6
pyrsistent==0.18.0
python-dateutil==2.8.2
PyYAML==6.0
ray==1.8.0
redis==4.0.1
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
scipy==1.7.2
setuptools-scm==6.3.2
six==1.16.0
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.7.0
tensorflow-estimator==2.7.0
tensorflow-io-gcs-filesystem==0.22.0
termcolor==1.1.0
tomli==1.2.2
tqdm==4.62.3
typing-extensions==4.0.0
urllib3==1.26.7
Werkzeug==2.0.2
wrapt==1.13.3
zipp==3.6.0
